Restructure ready for redux
