+++ src/pages/PetLostForm.css
@@ -1,0 +1,152 @@
+/* .submission-form-container {
+  padding-top: 50px;
+  text-align: center;
+  background: #1D3557;
+  color: white;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
+
+.submission-form-container h1 {
+  margin-top: 50px;
+  margin-bottom: 20px;
+  color: #fff;
+}
+
+.submission-form-container p {
+  max-width: 600px;
+  margin-bottom: 30px;
+}
+
+.submission-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  max-width: 600px;
+  background: #f5f6f7;
+  padding: 20px;
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  margin: 0 auto;
+  box-sizing: border-box;
+}
+
+.form-group {
+  width: 100%;
+  margin-bottom: 15px;
+  box-sizing: border-box;
+}
+
+.image-upload {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #fff;
+  border: 2px solid #e63946;
+  border-radius: 5px;
+  cursor: pointer;
+  padding: 10px;
+  width: 100%;
+  box-sizing: border-box;
+  color: black;
+  font-style: italic;
+}
+
+.image-upload label {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+}
+
+.image-upload input {
+  display: none;
+  color: black;
+}
+
+input[type="text"],
+input[type="email"],
+input[type="date"],
+textarea {
+  width: 100%;
+  padding: 10px;
+  border: 2px solid #e63946;
+  border-radius: 5px;
+  font-size: 16px;
+  box-sizing: border-box;
+}
+
+.image-upload-header {
+  margin-bottom: 5px;
+  text-align: left;
+}
+
+.image-upload-header h3 {
+  margin: 0 0 5px 0;
+  font-size: 16px;
+  color: #343a40;
+}
+
+.image-upload-header p {
+  margin: 2px;
+  font-size: 12px;
+  color: #6c757d;
+}
+
+.uploaded-image button {
+  margin-right: 10px;
+  margin-left: 10px;
+}
+
+textarea {
+  resize: none;
+}
+
+.confirmation-message {
+  margin-top: 20px;
+  margin-bottom: 20px;
+  width: 100%;
+  max-width: 600px;
+  display: flex;
+  padding-top: 10px;
+  padding-bottom: 10px;
+  background-color: #dff0d8;
+  color: #3c763d;
+  border: 1px solid #d6e9c6;
+  border-radius: 4px;
+}
+
+.disclaimer {
+  margin-top: 20px;
+  font-size: 12px;
+}
+
+.disclaimer a {
+  color: #007bff;
+  text-decoration: underline;
+}
+
+.image-upload span {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 10px;
+  border: none;
+  height: 35px;
+  text-align: center;
+  font-size: 14px;
+  font-style: normal;
+  background-color: #e63946;
+  border-color: #e63946;
+  color: white; 
+  width: 100px;
+}
+
+.image-upload span:hover {
+  background-color: white; 
+  color: #e63946; 
+}
+ */
