+++ src/pages/PetLostForm.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import SubmissionForm from '../components/forms/submissionForm/SubmissionForm';
+import Layout from "../components/layout/Layout";
+
+const PetLostForm = () => {
+  return (
+    <>
+      <Layout>
+      <SubmissionForm />
+      </Layout>
+    </>
+  );
+};
+
+export default PetLostForm;
